Be enum

+++ be/app/Observers/UserObserver.php
@@ -13,7 +13,7 @@ class UserObserver
      */
     public function updating(User $user): void
     {
-        if ($user->isDirty('profile_pic')){
+        if ($user->isDirty('profile_pic') && $user->getOriginal('profile_pic')){
             General::deleteFile($user->getOriginal('profile_pic'), 'user');
         }
     }
